refactor(wpapi): separate C FFI bindings into ffi.rs (close #45)
